Converted filters to objects
