--- examples/README.txt
@@ -1,6 +1,0 @@
-.. _general_examples:
-
-General examples
-================
-
-Introductory examples.
